prefactor(python-sdk): helpers for stream methods
Allows re-use in the RetryStreamResponse
